Refactor: make it easier to fetch a specific release manifest (#38)
This will be used by the provision tool to allow specific versions of each of the components to be installed if requested. There is a change to the behaviour of the core loop here; in some cases now the loop will log an error and continue instead of failing and terminating the loop. This seems like a good change. It's important that any component we install is verified, but if a previous component can't be verified then blocking updates forever feels like an unnecessary hill to die on.